+++ src/main.rs
@@ -103,19 +103,9 @@ async fn main() -> anyhow::Result<()> {
                 open_report,
                 num_runs,
                 title,
-                template_path,
             } => {
-                let custom_template = if let Some(ref template_path) = template_path {
-                    Some(std::fs::read_to_string(template_path).with_context(|| {
-                        format!("failed to read HTML template from {:?}", template_path)
-                    })?)
-                } else {
-                    None
-                };
-
                 let config = ReportConfig {
                     output_path,
-                    custom_template,
                     title,
                     num_runs,
                     open_report,